move CloudwatchTaskHandler to aws
